requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
